+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.6.4.3]
+- Hide Apple Pay option for non-safari browsers
+- Hide external payment site redirection note when Apple Pay selected
+
 ## [1.6.4.2]
 - Get the currency code dynamically for Apple Pay
 - Display warnings using NetSuite native approach when Apple Pay fails